requirements.txt
pathlib
streamlit
pillow